Implement start new game matcher
